requirements.txt
pytest-html
allure-pytest
pytest-xdis
pytest-orderding
pytest-rerunfailures